# backend/schemas/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateSessionRequest": {
        "additionalProperties": false,
        "properties": {
          "acceptance_criteria": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Acceptance Criteria",
            "type": "array"
          },
          "auto_merge": {
            "default": false,
            "title": "Auto Merge",
            "type": "boolean"
          },
          "base_ref": {
            "default": "HEAD",
            "title": "Base Ref",
            "type": "string"
          },
          "harness": {
            "minLength": 1,
            "title": "Harness",
            "type": "string"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "prompt": {
            "minLength": 1,
            "title": "Prompt",
            "type": "string"
          },
          "repo_path": {
            "format": "path",
            "title": "Repo Path",
            "type": "string"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "required": [
          "repo_path",
          "prompt",
          "harness"
        ],
        "title": "CreateSessionRequest",
        "type": "object"
      },
...
      "CreatedSessionResponse": {
        "properties": {
          "node": {
            "$ref": "#/components/schemas/NodeResponse"
          },
          "session": {
            "$ref": "#/components/schemas/SessionResponse"
          }
        },
        "required": [
          "session",
          "node"
        ],
        "title": "CreatedSessionResponse",
        "type": "object"
      },
...
      "DiffResponse": {
        "properties": {
          "patch": {
            "title": "Patch",
            "type": "string"
          }
        },
        "required": [
          "patch"
        ],
        "title": "DiffResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MergeResponse": {
        "properties": {
          "commit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Commit"
          },
          "conflicts": {
            "items": {
              "format": "path",
              "type": "string"
            },
            "title": "Conflicts",
            "type": "array"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "commit",
          "conflicts"
        ],
        "title": "MergeResponse",
        "type": "object"
      },
      "NodeResponse": {
        "properties": {
          "acceptance_criteria": {
            "items": {
              "type": "string"
            },
            "title": "Acceptance Criteria",
            "type": "array"
          },
          "base_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Ref"
          },
          "branch": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Branch"
          },
          "created_ms": {
            "title": "Created Ms",
            "type": "integer"
          },
          "estimated_effort": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Effort"
          },
          "harness": {
            "title": "Harness",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "prompt": {
            "title": "Prompt",
            "type": "string"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/NodeStatus"
          },
          "touches": {
            "items": {
              "type": "string"
            },
            "title": "Touches",
            "type": "array"
          },
          "updated_ms": {
            "title": "Updated Ms",
            "type": "integer"
          },
          "worktree_path": {
            "anyOf": [
              {
                "format": "path",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Worktree Path"
          }
        },
        "required": [
          "id",
          "session_id",
          "name",
          "prompt",
          "acceptance_criteria",
          "harness",
          "model",
          "touches",
          "estimated_effort",
          "worktree_path",
          "branch",
          "base_ref",
          "status",
          "created_ms",
          "updated_ms"
        ],
        "title": "NodeResponse",
        "type": "object"
      },
...
      "RunOutcomeResponse": {
        "properties": {
          "block_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Block Reason"
          },
          "commit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Commit"
          },
          "cost_complete": {
            "title": "Cost Complete",
            "type": "boolean"
          },
          "estimated_equivalent_cost_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Equivalent Cost Usd"
          },
          "merged": {
            "title": "Merged",
            "type": "boolean"
          },
          "node_id": {
            "title": "Node Id",
            "type": "string"
          },
          "node_status": {
            "$ref": "#/components/schemas/NodeStatus"
          },
          "permission_denials": {
            "title": "Permission Denials",
            "type": "integer"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "run_status": {
            "$ref": "#/components/schemas/RunState"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "tokens": {
            "$ref": "#/components/schemas/TokenCountsResponse"
          },
          "trusted": {
            "title": "Trusted",
            "type": "boolean"
          }
        },
        "required": [
          "session_id",
          "node_id",
          "run_id",
          "run_status",
          "node_status",
          "trusted",
          "permission_denials",
          "tokens",
          "estimated_equivalent_cost_usd",
          "cost_complete",
          "commit",
          "merged",
          "block_reason"
        ],
        "title": "RunOutcomeResponse",
        "type": "object"
      },
      "RunResponse": {
        "properties": {
          "attempt": {
            "title": "Attempt",
            "type": "integer"
          },
          "created_ms": {
            "title": "Created Ms",
            "type": "integer"
          },
          "event_count": {
            "title": "Event Count",
            "type": "integer"
          },
          "exit_code": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exit Code"
          },
          "finished_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished Ms"
          },
          "harness": {
            "title": "Harness",
            "type": "string"
          },
          "harness_session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Harness Session Id"
          },
          "harness_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Harness Version"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "node_id": {
            "title": "Node Id",
            "type": "string"
          },
          "permission_denial_count": {
            "title": "Permission Denial Count",
            "type": "integer"
          },
          "pid": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pid"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "started_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started Ms"
          },
          "status": {
            "$ref": "#/components/schemas/RunState"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          }
        },
        "required": [
          "id",
          "node_id",
          "session_id",
          "attempt",
          "status",
          "harness",
          "model",
          "pid",
          "harness_session_id",
          "harness_version",
          "started_ms",
          "finished_ms",
          "exit_code",
          "summary",
          "event_count",
          "permission_denial_count",
          "created_ms"
        ],
        "title": "RunResponse",
        "type": "object"
      },
      "RunState": {
        "description": "One execution of a node.\n\n``RUNNING`` is the only non-terminal member and the only one with no\ncounterpart in ``RunStatus``: a harness never *reports* \"still running\", it\nis the state of a row between :class:`~app.harnesses.events.RunStarted` and\n:class:`~app.harnesses.events.RunFinished`. A row still ``RUNNING`` after an\norchestrator restart is an orphan, and the scheduler resolves it to\n``INTERRUPTED`` rather than to a sixth state.",
        "enum": [
          "running",
          "success",
          "failed",
          "interrupted",
          "budget_exceeded"
        ],
        "title": "RunState",
        "type": "string"
      },
      "RunSummaryResponse": {
        "properties": {
          "cost_complete": {
            "title": "Cost Complete",
            "type": "boolean"
          },
          "estimated_equivalent_cost_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Equivalent Cost Usd"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "tokens": {
            "$ref": "#/components/schemas/TokenCountsResponse"
          },
          "trusted": {
            "title": "Trusted",
            "type": "boolean"
          }
        },
        "required": [
          "run_id",
          "trusted",
          "tokens",
          "estimated_equivalent_cost_usd",
          "cost_complete"
        ],
        "title": "RunSummaryResponse",
        "type": "object"
      },
      "SessionResponse": {
        "properties": {
          "auto_merge": {
            "title": "Auto Merge",
            "type": "boolean"
          },
          "created_ms": {
            "title": "Created Ms",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "integration_branch": {
            "title": "Integration Branch",
            "type": "string"
          },
          "repo_path": {
            "format": "path",
            "title": "Repo Path",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/SessionStatus"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_ms": {
            "title": "Updated Ms",
            "type": "integer"
          },
          "workspace_root": {
            "format": "path",
            "title": "Workspace Root",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "repo_path",
          "workspace_root",
          "integration_branch",
          "auto_merge",
          "status",
          "created_ms",
          "updated_ms"
        ],
        "title": "SessionResponse",
        "type": "object"
      },
      "SessionStatus": {
        "description": "One planning conversation plus its graph.",
        "enum": [
          "planning",
          "running",
          "paused",
          "done",
          "failed"
        ],
        "title": "SessionStatus",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "AgentHub",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/sessions": {
      "get": {
        "operationId": "list_sessions_api_sessions_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SessionResponse"
                  },
                  "title": "Response List Sessions Api Sessions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Sessions",
        "tags": [
          "sessions"
        ]
      },
      "post": {
        "operationId": "create_session_api_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Session",
        "tags": [
          "sessions"
        ]
      }
    },
    "/api/sessions/{session_id}": {
      "get": {
        "operationId": "get_session_api_sessions__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Session",
        "tags": [
          "sessions"
        ]
      }
    },
    "/api/sessions/{session_id}/approve": {
      "post": {
        "operationId": "approve_api_sessions__session_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MergeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approve",
        "tags": [
          "sessions"
        ]
      }
    },
    "/api/sessions/{session_id}/diff": {
      "get": {
        "operationId": "get_diff_api_sessions__session_id__diff_get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiffResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Diff",
        "tags": [
          "sessions"
        ]
      }
    },
    "/api/sessions/{session_id}/kill": {
      "post": {
        "operationId": "kill_run_api_sessions__session_id__kill_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Kill Run",
        "tags": [
          "sessions"
        ]
      }
    },
    "/api/sessions/{session_id}/node": {
      "get": {
        "operationId": "get_node_api_sessions__session_id__node_get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NodeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Node",
        "tags": [
          "sessions"
        ]
      }
    },
...
    "/api/sessions/{session_id}/retry": {
      "post": {
        "operationId": "retry_run_api_sessions__session_id__retry_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunOutcomeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Retry Run",
        "tags": [
          "sessions"
        ]
      }
    },
    "/api/sessions/{session_id}/runs": {
      "get": {
        "operationId": "list_runs_api_sessions__session_id__runs_get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RunResponse"
                  },
                  "title": "Response List Runs Api Sessions  Session Id  Runs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Runs",
        "tags": [
          "sessions"
        ]
      },
      "post": {
        "operationId": "start_run_api_sessions__session_id__runs_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunOutcomeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Run",
        "tags": [
          "sessions"
        ]
      }
    },
    "/api/sessions/{session_id}/runs/{run_id}/events": {
      "get": {
        "operationId": "list_run_events_api_sessions__session_id__runs__run_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Canonical persisted AgentEvent array"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Run Events",
        "tags": [
          "sessions"
        ]
      }
    },
    "/api/sessions/{session_id}/runs/{run_id}/summary": {
      "get": {
        "operationId": "get_run_summary_api_sessions__session_id__runs__run_id__summary_get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunSummaryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Run Summary",
        "tags": [
          "sessions"
        ]
      }
    },
...
  }
}